requirements.txt
flask
flask_session
requests
python-dotenv
gunicorn
